[ADD] Allowed PotentialLayouts to work on the highest level
Add a AddPotentialLayouts function to avoid duplicate code as I need to add them in the intialize function as well. This isn't enough to get it to work as the AddBlock function couldn't look for PotentialLayouts in the Canvas.

To fix this I added a FindBlock function with different arguments which will either create a PotentialLayout, find a block from the Linked Blocks, or returns with an error message. This decision was made to not clutter the AddData function as the only difference was the pointer for the previousBlock that needed to be different.