[CHANGE] the moving up section of an image portion
With the way it was working before there was a scenario where not everything that should have been erased was leaving duplicate parts in the image which doesn't look good. There was also the possiblity of it moving the border out of the image bounds so I added a check to prevent it from doing so.